% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_histograms.R
\name{categorical_histograms}
\alias{categorical_histograms}
\title{Plot exploratory visual of histograms/boxplot by feature}
\usage{
categorical_histograms(data, features, plot_width = 600, plot_height = 400)
}
\arguments{
\item{data}{Data to plot}

\item{features}{Which features from \code{data} to plot as a character vector}

\item{plot_width}{Width of the plot produced in pixels}

\item{plot_height}{Height of the plot produced in pixels}
}
\value{
ggplot2 object
}
\description{
Will plot histograms for numerical features, box plots for categorical.
}
\examples{
categorical_histograms(mtcars, features=c("mpg","cyl"), plot_width=600, plot_height=400)
}
