% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaziReda-package.R
\docType{package}
\name{eaziReda-package}
\alias{eaziReda}
\alias{eaziReda-package}
\title{eaziReda: A Quick And Easy Way To Do EDA And Preprocessing}
\description{
Almost every data analysis project involves the process of doing some exploratory data analysis(EDA) and data preprocessing. 
  Usually they serve as a very crucial and inevitable step in a data analysis workflow. 
  There are some very common tasks in EDA, which can include checking missing values, detecting outliers, ploting correlation plots between features
  and ploting histograms/bar plots for each individual features.
  Typically these steps are followed by some preprocesing like imputation and dealing with outliers. 
  All of those steps together may require lots of coding effort and can be repeated for several projects. 
  To solve this issue, we designed this R package eaziReda that wraps all of those lines of code into four convenient 
  functions that will allow you to quickly and easily carry out EDA along with some simple preprocessing using just four lines of code!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBC-MDS/eaziReda}
  \item Report bugs at \url{https://github.com/UBC-MDS/eaziReda/issues}
}

}
\author{
\strong{Maintainer}: Dustin Andrews \email{dandrew9@student.ubc.ca}

Authors:
\itemize{
  \item Vignesh Rajakumar
  \item Arash Shamseddini
  \item Yuyan Guo
}

}
\keyword{internal}
