% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{histograms}
\alias{histograms}
\title{Plot exploratory visual of histograms/bar chart by feature}
\usage{
histograms(data, features, num_cols = 2)
}
\arguments{
\item{data}{data.frame like object to plot histograms/bar charts}

\item{features}{Which features from \code{data} to plot as a character vector}

\item{num_cols}{Number of columns in facetted chart produced}
}
\value{
ggplot2 object
}
\description{
Will plot histograms for numerical features, bar charts for categorical.
}
\examples{
histograms(mtcars, features=c("mpg","cyl"))
}
