% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_detect.R
\name{missing_detect}
\alias{missing_detect}
\title{Check missing values in dataframe}
\usage{
missing_detect(data)
}
\arguments{
\item{data}{Input dataframe to checked with missing values}
}
\value{
dataframe containing two columns: the number of missing values and
the percentage of missing values for each column
}
\description{
Return the number/percentage of missing values for each column in the dataframe
}
\examples{
missing_detect(mtcars)
}
