% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_impute.R
\name{missing_impute}
\alias{missing_impute}
\title{Return the imputed version of data based on the methods selected}
\usage{
missing_impute(data, method_num = "mean", method_non_num = "most_frequent")
}
\arguments{
\item{data}{Input dataframe to apply imputation}

\item{method_num}{The method used for imputing numerical missing values}

\item{method_non_num}{The method used for imputing non-numerical missing values}
}
\value{
dataframe with imputations applied
}
\description{
Return the imputed version of data based on the methods selected
}
\examples{
missing_impute(mtcars)
}
