% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detect.R
\name{outliers_detect}
\alias{outliers_detect}
\title{Detect outliers in the given vector}
\usage{
outliers_detect(s, method = "zscore")
}
\arguments{
\item{s}{vector that contains the data to detect outliers in}

\item{method}{method to detect outliers. One of 'zscore', 'iqr', 'iforest'}
}
\value{
boolean vecotr with indices makerd true for the outliers
}
\description{
Detect outliers in the given vector
}
\examples{
outliers_detect(c(1,2,1,2,100))
}
