% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detect.R
\name{outliers_detect_iforest}
\alias{outliers_detect_iforest}
\title{Detects outliers in a data.frame/tibble using isolation forests}
\usage{
outliers_detect_iforest(s, threshold = 0.5)
}
\arguments{
\item{s}{numeric vector for which the outliers need to be found}

\item{threshold}{threshold for considering an outlier}
}
\value{
boolean vector with indices makerd true for the outliers
}
\description{
Detects outliers in a data.frame/tibble using isolation forests
}
\examples{
outliers_detect_iforest(c(1,1,1,1,1,1,1,1,1,1,1e14))
}
