% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detect.R
\name{outliers_detect_iqr}
\alias{outliers_detect_iqr}
\title{Detects outliers in a data.frame/tibble using inter-quantile ranges}
\usage{
outliers_detect_iqr(s, threshold = 1.5)
}
\arguments{
\item{s}{numeric vector for which the outliers need to be found}

\item{threshold}{iqr factor used for outliers}
}
\value{
boolean vector with indices makerd true for the outliers
}
\description{
Detects outliers in a data.frame/tibble using inter-quantile ranges
}
\examples{
outliers_detect_iqr(c(1,1,1,1,1,1,1,1,1,1,1e14))
}
