% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detect.R
\name{outliers_detect_zscore}
\alias{outliers_detect_zscore}
\title{Detects outliers in a data.frame/tibble using zscores}
\usage{
outliers_detect_zscore(s, threshold = 3)
}
\arguments{
\item{s}{numeric vector for which the outliers need to be found}

\item{threshold}{zscore threshold used for outliers}
}
\value{
boolean vector with indices makerd true for the outliers
}
\description{
Detects outliers in a data.frame/tibble using zscores
}
\examples{
outliers_detect_zscore(c(1,1,1,1,1,1,1,1,1,1,1e14))
}
