% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detect.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(s, outliers)
}
\arguments{
\item{s}{vecotr with the data to delete outliers from}

\item{outliers}{boolean vector with outliers marked true}
}
\value{
vector with outliers deleted
}
\description{
Remove outliers
}
\examples{
remove_outliers(c(1,2,100), c(F,F,T))
}
